make AUDIO/VIDEO PROPERTIES constant
