+++ lib/constants/audio-properties.js
@@ -1,0 +1,10 @@
+// constants
+var AUDIO_PROPERTIES = [
+  'audioobjecttype',
+  'channelcount',
+  'samplerate',
+  'samplingfrequencyindex',
+  'samplesize'
+];
+
+module.exports = AUDIO_PROPERTIES;
