+++ lib/constants/video-properties.js
@@ -1,0 +1,11 @@
+var VIDEO_PROPERTIES = [
+  'width',
+  'height',
+  'profileIdc',
+  'levelIdc',
+  'profileCompatibility',
+  'sarRatio'
+];
+
+
+module.exports = VIDEO_PROPERTIES;
