+++ lib/mp4/transmuxer.js
@@ -22,25 +22,8 @@ var H264Stream = require('../codecs/h264').H264Stream;
 var AacStream = require('../aac');
 var isLikelyAacData = require('../aac/utils').isLikelyAacData;
 var ONE_SECOND_IN_TS = require('../utils/clock').ONE_SECOND_IN_TS;
-
-
-// constants
-var AUDIO_PROPERTIES = [
-  'audioobjecttype',
-  'channelcount',
-  'samplerate',
-  'samplingfrequencyindex',
-  'samplesize'
-];
-
-var VIDEO_PROPERTIES = [
-  'width',
-  'height',
-  'profileIdc',
-  'levelIdc',
-  'profileCompatibility',
-  'sarRatio'
-];
+var AUDIO_PROPERTIES = require('../constants/audio-properties.js');
+var VIDEO_PROPERTIES = require('../constants/video-properties.js');
 
 // object types
 var VideoSegmentStream, AudioSegmentStream, Transmuxer, CoalesceStream;
