+++ lib/partial/audio-segment-stream.js
@@ -5,15 +5,7 @@ var mp4 = require('../mp4/mp4-generator.js');
 var audioFrameUtils = require('../mp4/audio-frame-utils');
 var trackInfo = require('../mp4/track-decode-info.js');
 var ONE_SECOND_IN_TS = require('../utils/clock').ONE_SECOND_IN_TS;
-
-// constants
-var AUDIO_PROPERTIES = [
-  'audioobjecttype',
-  'channelcount',
-  'samplerate',
-  'samplingfrequencyindex',
-  'samplesize'
-];
+var AUDIO_PROPERTIES = require('../constants/audio-properties.js');
 
 /**
  * Constructs a single-track, ISO BMFF media segment from AAC data
