+++ lib/partial/transmuxer.js
@@ -263,6 +263,8 @@ var Transmuxer = function(options) {
     pipeline = null,
     hasFlushed = true;
 
+  options = options || {};
+
   Transmuxer.prototype.init.call(this);
   options.baseMediaDecodeTime = options.baseMediaDecodeTime || 0;
