+++ lib/partial/video-segment-stream.js
@@ -13,14 +13,7 @@ var Stream = require('../utils/stream.js');
 var mp4 = require('../mp4/mp4-generator.js');
 var trackInfo = require('../mp4/track-decode-info.js');
 var frameUtils = require('../mp4/frame-utils');
-
-var VIDEO_PROPERTIES = [
-  'width',
-  'height',
-  'profileIdc',
-  'levelIdc',
-  'profileCompatibility'
-];
+var VIDEO_PROPERTIES = require('../constants/video-properties.js');
 
 var VideoSegmentStream = function(track, options) {
   var
